auto-incremented values for the id field

+++ src/main/java/edu/nyu/nsg2057/webscraper/repository/MonitorRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MonitorRepository extends MongoRepository<Monitor, Integer > {
+public interface MonitorRepository extends MongoRepository<Monitor, Long> {
 
 }
